-- change the type of variable for holding the size of free memory to allow for larger memory buffers in tafel.cpp
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@682 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ src/tafel.cpp
@@ -2447,7 +2447,7 @@ int32_t ibis::tafel::reserveBuffer(uint32_t maxr) {
 	}
 	if (rowsize > 0) {
 	    rowsize += rowsize;
-	    uint32_t tmp = ibis::fileManager::bytesFree();
+	    size_t tmp = ibis::fileManager::bytesFree();
 	    if (tmp < 10000000) tmp = 10000000;
 	    tmp = tmp / rowsize;
 	    if (tmp < maxr) {
